[BEAM-12572] Move tests to integration tests files
